Change Flow Control cursor to be compliant with OAI-PMH guidelines
The cursor is defined as "a count of the number of elements of the
complete list thus far returned (i.e. cursor starts at 0)."
(http://www.openarchives.org/OAI/openarchivesprotocol.html#FlowControl)
 
Previously a simple counter starting from 1 was used as cursor for Flow
Control.

This commit changes the cursor to be compliant with the OAI-PMH
guidelines